services/device_service/requirements.txt
alembic==1.9.0
anyio==3.6.2
certifi==2022.12.7
click==8.1.3
dnspython==2.2.1
email-validator==1.3.0
fastapi==0.88.0
fastapi-utils==0.2.1
greenlet==2.0.1
h11==0.14.0
httpcore==0.16.2
httptools==0.5.0
httpx==0.23.1
idna==3.4
importlib-metadata==5.1.0
importlib-resources==5.10.1
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.1
orjson==3.8.3
psycopg2==2.9.5
pydantic==1.10.2
python-dotenv==0.21.0
python-multipart==0.0.5
PyYAML==6.0
rfc3986==1.5.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.45
SQLAlchemy-Utils==0.38.3
starlette==0.22.0
typing_extensions==4.4.0
ujson==5.6.0
uvicorn==0.20.0
uvloop==0.17.0
watchfiles==0.18.1
websockets==10.4
zipp==3.11.0
